Make method to convert a board into ascii strings
Method 'board_to_ascii' should take two arguments: a board and a
hash that maps the symbols on the board to ascii characters. The
output should be an array of strings. For a 3x3 board, there will
be three strings, each representing three spaces on the board. The
spaces will be separated by a vertical pipe (|) character.