Move cloning of player list into #initialize
The Game class constructor will have access to the finalized game
configuration. At initialization is when the players list should be
cloned. Remove this action from the #set_players method and place it
into the #initialize method.

Player instantiation tests will break. Be sure to remove these as they
are no longer needed.