Add method #prompt_player_order
Method #prompt_player_order should take a Hash as an argument,
containing player symbols mapped to player marks, and return an array as
a result. The return array should contain a sorted list of the symbols from
the Hash. The first position in the array should be given to whichever
symbol was mapped to the character 'X'.