Make #generate_solutions work for board size 4
Method #generate_solutions should take no arguments and return no value.
The method should fill the 'solutions' attribute with a two-dimensional
array holding integers. Each integer array should represent a list of
board spaces which are a winning solution for the board.

This method should create a full list of horizontal, vertical, and
diagonal solutions for board sizes of 3 and 4. The length of each
solution array should be equal to the size of the board.