Add method #prompt_board_size to get board size
Method #prompt_board_size should take no arguments and return an integer
representing the desired size of the board. The console prompter should
be set with valid_inputs of '3x3' and '4x4' for the accepted board
sizes. A request should be issued with a clear message requesting board
size from the user. The input should then be converted to an integer
appropriate to construct a board of the requested dimensions.