Make #prompt_player_mark use an array argument
Method #prompt_player_mark should receive an array as an argument, and
return an integer result. The array passed to the method should be a
list of the BLANK spaces on the game board. The resulting integer should
be one of these indices as designated by the player/user.

Each value in the array argument should be incremented for the sake of
presentation and converted to a String form. These values should be
stored in the 'valid_input' attribute of the console's Prompter object.
This part is necessary to ensure that input to the prompter is matched
to a String.

When input is supplied to and returned from the prompter, it is
converted back to an integer and decremented to account for the
original increment for presentation.

Use of the method #prompt_player_mark by the Human class should account
for the change in arguments.