Indicate winner using mark instead of number
With the way that the Configuration class handles player ordering, first
and second player are indistinguishable. This is also true for during
gameplay. It is clearer and more appropriate to make the distinction
using the marks assigned to the players.

To accomplish this, Game method #display_game_results should pass the
winning symbol to the console method #display_game_winner, which will
display a message using the ascii character associated with that symbol.